Refactor args validation and add options in README.
