+++ client/config.yaml
@@ -14,7 +14,7 @@
 
 gcp_settings:
   # GCP project number.
-  project_number: 'XXXXXXXXXX'
+  project_number: 'translation-test-external'
 
   # Google Cloud Storage bucket, example: 'exemplary_client_test_external'
   gcs_bucket: 'exemplary_client_test_external'
